more refactors for modularity
